# perf: Improve the performance of log.LTSVReader