benchmark result

```
goos: darwin
goarch: arm64
pkg: github.com/haijima/stool/internal/log
              │     Old     │                 New                 │
              │   sec/op    │   sec/op     vs base                │
LTSVReader-10   523.2m ± 1%   251.8m ± 1%  -51.87% (p=0.000 n=10)

              │      Old      │                 New                  │
              │     B/op      │     B/op      vs base                │
LTSVReader-10   450.28Mi ± 0%   88.96Mi ± 0%  -80.24% (p=0.000 n=10)

              │     Old     │                 New                 │
              │  allocs/op  │  allocs/op   vs base                │
LTSVReader-10   6.450M ± 0%   1.291M ± 0%  -79.98% (p=0.000 n=10)
```